vickrey_auction_GUI addition
Added new folder for VA_gui.

Working on grpc communication between client(gui on tablet) and server(gui on main computer for experimenter).

Client will be able to send bids to the server and subscribes to a stream which announces the results of each 2 minute auction(WIP).

Server receives bids from client and announces results to client(WIP). Server will keep track of robo-bidders, run the VA and log the results (WIP).